# RedWolf/API/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "RedWolf API",
    "version": "v0.1"
  },
  "paths": {
...
    "/api/indicators/files/{id}": {
      "get": {
        "tags": [
          "IndicatorApi"
        ],
        "operationId": "GetFileIndicator",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FileIndicator"
            }
          }
        }
      }
    },
...
    "/api/profiles": {
      "get": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "GetProfiles",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Profile"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "CreateProfile",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "EditProfile",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        }
      }
    },
    "/api/profiles/{id}": {
      "get": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "GetProfile",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "DeleteProfile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
    "/api/profiles/http": {
      "get": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "GetHttpProfiles",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/HttpProfile"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "CreateHttpProfile",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HttpProfile"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpProfile"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "EditHttpProfile",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/HttpProfile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        }
      }
    },
    "/api/profiles/http/{id}": {
      "get": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "GetHttpProfile",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/HttpProfile"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProfileApi"
        ],
        "operationId": "DeleteHttpProfile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/themes": {
      "get": {
        "tags": [
          "ThemeApi"
        ],
        "operationId": "GetThemes",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Theme"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ThemeApi"
        ],
        "operationId": "CreateTheme",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Theme"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Theme"
            }
          }
        }
      },
      "put": {
        "tags": [
          "ThemeApi"
        ],
        "operationId": "EditTheme",
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Theme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Theme"
            }
          }
        }
      }
    },
    "/api/themes/{id}": {
      "get": {
        "tags": [
          "ThemeApi"
        ],
        "operationId": "GetTheme",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Theme"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ThemeApi"
        ],
        "operationId": "DeleteTheme",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "CodeMirrorTheme": {
      "enum": [
        "default",
        "night"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "CodeMirrorTheme",
        "modelAsString": false
      }
    },
    "Theme": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "backgroundColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "backgroundTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "primaryHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "secondaryHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "terminalBorderColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "navbarColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "sidebarColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputDisabledColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputTextColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "inputHighlightColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "textLinksColor": {
          "pattern": "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$",
          "type": "string"
        },
        "codeMirrorTheme": {
          "$ref": "#/definitions/CodeMirrorTheme"
        }
      }
    },
...
    "ProfileType": {
      "enum": [
        "HTTP",
        "Bridge"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "ProfileType",
        "modelAsString": false
      }
    },
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/ProfileType"
        },
        "messageTransform": {
          "type": "string"
        }
      }
    },
...
    "IndicatorType": {
      "enum": [
        "FileIndicator",
        "NetworkIndicator",
        "TargetIndicator"
      ],
      "type": "string",
      "x-ms-enum": {
        "name": "IndicatorType",
        "modelAsString": false
      }
    },
...
    "FileIndicator": {
      "type": "object",
      "properties": {
        "fileName": {
          "type": "string"
        },
        "filePath": {
          "type": "string"
        },
        "shA2": {
          "type": "string"
        },
        "shA1": {
          "type": "string"
        },
        "mD5": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "$ref": "#/definitions/IndicatorType"
        }
      }
    },
...
    "HttpProfileHeader": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    },
    "HttpProfile": {
      "type": "object",
      "properties": {
        "httpUrls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "httpRequestHeaders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/HttpProfileHeader"
          }
        },
        "httpResponseHeaders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/HttpProfileHeader"
          }
        },
        "httpPostRequest": {
          "type": "string"
        },
        "httpGetResponse": {
          "type": "string"
        },
        "httpPostResponse": {
          "type": "string"
        },
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "$ref": "#/definitions/ProfileType"
        },
        "messageTransform": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}